+++ src/main/java/com/epam/hotel/action/impl/DeleteOrderRoomDetailAction.java
@@ -2,39 +2,35 @@ package com.epam.hotel.action.impl;
 
 import com.epam.hotel.action.Action;
 import com.epam.hotel.dao.OrderRoomDetailDAO;
-import com.epam.hotel.dao.RoomDAO;
 import com.epam.hotel.dao.impl.OrderRoomDetailDAOImpl;
-import com.epam.hotel.dao.impl.RoomDAOImpl;
-import com.epam.hotel.entity.Room;
-import com.epam.hotel.entity.role.Role;
+import com.epam.hotel.validation.NumericValidation;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import static com.epam.hotel.action.impl.ActionConstant.*;
+import static com.epam.hotel.util.constant.ActionConstant.*;
+import static com.epam.hotel.util.constant.ErrorConstant.ERROR_INVALID_DATA;
 
 public class DeleteOrderRoomDetailAction implements Action {
     @Override
-    public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        long orderDetailId = Long.parseLong(request.getParameter(ORDER_ROOM_DETAIL_ID));
-        long orderMainId = Long.parseLong(request.getParameter(ORDER_MAIN_ID));
-        long roomId = Long.parseLong(request.getParameter(ActionConstant.ROOM_ID));
+    public void execute(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
-        OrderRoomDetailDAO orderRoomDetailDAO = new OrderRoomDetailDAOImpl();
-        RoomDAO roomDAO = new RoomDAOImpl();
-        Room room = roomDAO.getOneById(roomId);
-        room.setAvailability(Boolean.TRUE);
-        roomDAO.updateOneById(roomId, room);
-        orderRoomDetailDAO.deleteOneById(orderDetailId);
+        if (NumericValidation.isNumeric(request.getParameter(ORDER_ROOM_DETAIL_ID))
+                && NumericValidation.isNumeric(request.getParameter(ORDER_MAIN_ID))) {
 
-        request.setAttribute(ORDER_MAIN_ID, orderMainId);
+            long orderDetailId = Long.parseLong(request.getParameter(ORDER_ROOM_DETAIL_ID));
+            long orderMainId = Long.parseLong(request.getParameter(ORDER_MAIN_ID));
 
-        if (request.getSession().getAttribute(ROLE) == Role.ADMIN) {
-            request.getRequestDispatcher(ActionConstant.SHOW_ORDER_ROOM_DETAIL_URL).forward(request, response);
-            return;
+            OrderRoomDetailDAO orderRoomDetailDAO = new OrderRoomDetailDAOImpl();
+            orderRoomDetailDAO.deleteOneById(orderDetailId);
+
+            request.getSession().setAttribute(ORDER_MAIN_ID, orderMainId);
+            response.sendRedirect(SHOW_ORDER_ROOM_DETAIL_JSP);
+
+        } else {
+            request.getSession().setAttribute(MESSAGE, ERROR_INVALID_DATA);
+            response.sendRedirect(ERROR_JSP);
         }
-        request.getRequestDispatcher(ActionConstant.SHOW_ORDER_ROOM_DETAIL_URL).forward(request, response);
     }
 }
